Move Secure Session details to private headers
The TODO says to "probably move this to private headers" and this
is exactly what we're doing here. All of these definitions are
implementation details of Secure Session and do not belong to the
public Themis API. Move them to appropriate internal headers.
Also, drop all the swathes of commented unused code from there.
